requirements.txt
streamlit==1.43.2
pandas==2.2.3
openpyxl==3.1.5
XlsxWriter==3.2.2
plotly~=6.0.1
streamlit-aggrid==0.3.3
fpdf==1.7.2
kaleido==0.2.1
matplotlib~=3.9.4
pillow~=11.1.0